Split out template engine into “x-template.js”.
Conceptually “XElement” and “XTemplate” do very different things. The
former is all about defining _element interfaces_ and the latter is all
about turning interpolated markup into dom.

Over time, even our PR diffs have highlited this separation of concerns.
Often a PR _either_ changes XElement _or_ XTemplate — rarely both.
