+++ package.json
@@ -7,9 +7,11 @@
   "type": "module",
   "exports": {
     "./x-element.js": "./x-element.js",
-    "./x-element.d.ts": "./x-element.d.ts",
+    "./x-template.js": "./x-template.js",
     "./etc/ready.js": "./etc/ready.js",
-    "./etc/ready.d.ts": "./etc/ready.d.ts"
+    "./ts/x-element.d.ts": "./ts/x-element.d.ts",
+    "./ts/x-template.d.ts": "./ts/x-template.d.ts",
+    "./ts/etc/ready.d.ts": "./ts/etc/ready.d.ts"
   },
   "publishConfig": {
     "access": "public",
@@ -25,23 +27,23 @@
   },
   "files": [
     "/x-element.js",
-    "/x-element.d.ts",
-    "/x-element.d.ts.map",
+    "/x-template.js",
     "/demo",
     "/test",
-    "/etc"
+    "/etc",
+    "/ts"
   ],
   "devDependencies": {
-    "@netflix/eslint-config": "^3.0.0",
-    "eslint-plugin-jsdoc": "^50.3.1",
-    "eslint": "^9.12.0",
-    "puppeteer": "^23.5.3",
-    "tap-parser": "^18.0.0",
-    "typescript": "^5.6.3"
+    "@netflix/eslint-config": "3.0.0",
+    "eslint-plugin-jsdoc": "50.6.0",
+    "eslint": "9.16.0",
+    "puppeteer": "23.10.1",
+    "tap-parser": "18.0.0",
+    "typescript": "5.7.2"
   },
   "engines": {
-    "node": "20.18.0",
-    "npm": "10.8.2"
+    "node": "22.11.0",
+    "npm": "10.9.0"
   },
   "contributors": [
     {
