+++ ts/etc/ready.d.ts
