+++ ts/x-element.d.ts
@@ -147,55 +147,55 @@ export default class XElement extends HTMLElement {
     static template(html: Function, engine: {
         [key: string]: Function;
     }): (properties: object, host: HTMLElement) => any;
-    static "__#1@#analyzeConstructor"(constructor: any): void;
-    static "__#1@#validateProperties"(constructor: any, properties: any, entries: any): void;
-    static "__#1@#validateProperty"(constructor: any, key: any, property: any): void;
-    static "__#1@#validatePropertyAttribute"(constructor: any, key: any, property: any, attribute: any): void;
-    static "__#1@#propertyIsCyclic"(property: any, inputMap: any, seen?: Set<any>): boolean;
-    static "__#1@#validateListeners"(constructor: any, listeners: any, entries: any): void;
-    static "__#1@#mutateProperty"(constructor: any, propertyMap: any, key: any, property: any): void;
-    static "__#1@#addPropertyInitial"(constructor: any, property: any): void;
-    static "__#1@#addPropertyDefault"(constructor: any, property: any): void;
-    static "__#1@#addPropertySync"(constructor: any, property: any): void;
-    static "__#1@#addPropertyReflect"(constructor: any, property: any): void;
-    static "__#1@#addPropertyCompute"(constructor: any, property: any): void;
-    static "__#1@#addPropertyObserve"(constructor: any, property: any): void;
-    static "__#1@#constructHost"(host: any): void;
-    static "__#1@#createInternal"(host: any): any;
-    static "__#1@#createProperties"(host: any): any;
-    static "__#1@#connectHost"(host: any): void;
-    static "__#1@#disconnectHost"(host: any): void;
-    static "__#1@#initializeHost"(host: any): boolean;
-    static "__#1@#upgradeOwnProperties"(host: any): void;
-    static "__#1@#getPreUpgradePropertyValue"(host: any, property: any): {
+    static "__#2@#analyzeConstructor"(constructor: any): void;
+    static "__#2@#validateProperties"(constructor: any, properties: any, entries: any): void;
+    static "__#2@#validateProperty"(constructor: any, key: any, property: any): void;
+    static "__#2@#validatePropertyAttribute"(constructor: any, key: any, property: any, attribute: any): void;
+    static "__#2@#propertyIsCyclic"(property: any, inputMap: any, seen?: Set<any>): boolean;
+    static "__#2@#validateListeners"(constructor: any, listeners: any, entries: any): void;
+    static "__#2@#mutateProperty"(constructor: any, propertyMap: any, key: any, property: any): void;
+    static "__#2@#addPropertyInitial"(constructor: any, property: any): void;
+    static "__#2@#addPropertyDefault"(constructor: any, property: any): void;
+    static "__#2@#addPropertySync"(constructor: any, property: any): void;
+    static "__#2@#addPropertyReflect"(constructor: any, property: any): void;
+    static "__#2@#addPropertyCompute"(constructor: any, property: any): void;
+    static "__#2@#addPropertyObserve"(constructor: any, property: any): void;
+    static "__#2@#constructHost"(host: any): void;
+    static "__#2@#createInternal"(host: any): any;
+    static "__#2@#createProperties"(host: any): any;
+    static "__#2@#connectHost"(host: any): void;
+    static "__#2@#disconnectHost"(host: any): void;
+    static "__#2@#initializeHost"(host: any): boolean;
+    static "__#2@#upgradeOwnProperties"(host: any): void;
+    static "__#2@#getPreUpgradePropertyValue"(host: any, property: any): {
         value: any;
         found: boolean;
     };
-    static "__#1@#initializeProperty"(host: any, property: any): void;
-    static "__#1@#addListener"(host: any, element: any, type: any, callback: any, options: any): void;
-    static "__#1@#addListeners"(host: any): void;
-    static "__#1@#removeListener"(host: any, element: any, type: any, callback: any, options: any): void;
-    static "__#1@#removeListeners"(host: any): void;
-    static "__#1@#getListener"(host: any, listener: any): any;
-    static "__#1@#updateHost"(host: any): void;
-    static "__#1@#toPathString"(host: any): string;
-    static "__#1@#invalidateProperty"(host: any, property: any): Promise<void>;
-    static "__#1@#getPropertyValue"(host: any, property: any): any;
-    static "__#1@#validatePropertyValue"(host: any, property: any, value: any): void;
-    static "__#1@#setPropertyValue"(host: any, property: any, value: any): void;
-    static "__#1@#serializeProperty"(host: any, property: any, value: any): any;
-    static "__#1@#deserializeProperty"(host: any, property: any, value: any): any;
-    static "__#1@#propertyHasAttribute"(property: any): boolean;
-    static "__#1@#getTypeName"(value: any): any;
-    static "__#1@#notNullish"(value: any): boolean;
-    static "__#1@#typeIsWrong"(type: any, value: any): boolean;
-    static "__#1@#camelToKebab"(camel: any): any;
-    static "__#1@#constructors": WeakMap<WeakKey, any>;
-    static "__#1@#hosts": WeakMap<WeakKey, any>;
-    static "__#1@#propertyKeys": Set<string>;
-    static "__#1@#serializableTypes": Set<StringConstructor | BooleanConstructor | NumberConstructor>;
-    static "__#1@#caseMap": Map<any, any>;
-    static "__#1@#prototypeInterface": Set<string>;
+    static "__#2@#initializeProperty"(host: any, property: any): void;
+    static "__#2@#addListener"(host: any, element: any, type: any, callback: any, options: any): void;
+    static "__#2@#addListeners"(host: any): void;
+    static "__#2@#removeListener"(host: any, element: any, type: any, callback: any, options: any): void;
+    static "__#2@#removeListeners"(host: any): void;
+    static "__#2@#getListener"(host: any, listener: any): any;
+    static "__#2@#updateHost"(host: any): void;
+    static "__#2@#toPathString"(host: any): string;
+    static "__#2@#invalidateProperty"(host: any, property: any): Promise<void>;
+    static "__#2@#getPropertyValue"(host: any, property: any): any;
+    static "__#2@#validatePropertyValue"(host: any, property: any, value: any): void;
+    static "__#2@#setPropertyValue"(host: any, property: any, value: any): void;
+    static "__#2@#serializeProperty"(host: any, property: any, value: any): any;
+    static "__#2@#deserializeProperty"(host: any, property: any, value: any): any;
+    static "__#2@#propertyHasAttribute"(property: any): boolean;
+    static "__#2@#getTypeName"(value: any): any;
+    static "__#2@#notNullish"(value: any): boolean;
+    static "__#2@#typeIsWrong"(type: any, value: any): boolean;
+    static "__#2@#camelToKebab"(camel: any): any;
+    static "__#2@#constructors": WeakMap<WeakKey, any>;
+    static "__#2@#hosts": WeakMap<WeakKey, any>;
+    static "__#2@#propertyKeys": Set<string>;
+    static "__#2@#serializableTypes": Set<StringConstructor | BooleanConstructor | NumberConstructor>;
+    static "__#2@#caseMap": Map<any, any>;
+    static "__#2@#prototypeInterface": Set<string>;
     /**
      * Extends HTMLElement.prototype.connectedCallback.
      */
@@ -254,6 +254,6 @@ export default class XElement extends HTMLElement {
      * must get read-only properties directly from the host.
      * @returns {object}
      */
-    get internal(): any;
+    get internal(): object;
 }
 //# sourceMappingURL=x-element.d.ts.map
