+++ ts/x-template.d.ts
@@ -1,0 +1,20 @@
+export const render: any;
+export const html: any;
+export const svg: any;
+export const map: any;
+export const live: any;
+export const unsafeHTML: any;
+export const unsafeSVG: any;
+export const ifDefined: any;
+export const nullish: any;
+export const repeat: any;
+export const asyncAppend: any;
+export const asyncReplace: any;
+export const cache: any;
+export const classMap: any;
+export const directive: any;
+export const guard: any;
+export const styleMap: any;
+export const templateContent: any;
+export const until: any;
+//# sourceMappingURL=x-template.d.ts.map
