+++ tsconfig.json
@@ -1,10 +1,15 @@
 {
   "files": [
     "./x-element.js",
+    "./x-template.js",
     "./etc/ready.js"
   ],
   "compilerOptions": {
     "target": "ES2023",
+    // Beyond just organization, a “bug” in how JSDocs works causes a cyclic
+    //  import to be detected by TypeScript when letting generated “*.d.ts”
+    //  sit next to files.
+    "outDir": "ts",
     "allowJs": true,
     "checkJs": true,
     "declaration": true,
